refactor objects into own classes
