+++ redditapi/src/main/java/com/worker8/redditapi/model/t1_comment/RedditReply.kt
@@ -1,23 +1,22 @@
 package com.worker8.redditapi.model.t1_comment
 
 import com.google.gson.annotations.SerializedName
-import com.worker8.redditapi.model.listing.RedditCommentDataType
 import com.worker8.redditapi.ofType
 
 sealed class RedditReply {
-    class T1_RedditObject(@SerializedName("data") val value: RedditCommentDataType.RedditCommentData = RedditCommentDataType.RedditCommentData(), @SerializedName("kind") val kind: String = "") : RedditReply()
-    class TM_RedditObject(@SerializedName("data") val value: RedditCommentDataType.TMore = RedditCommentDataType.TMore(), @SerializedName("kind") val kind: String = "") : RedditReply()
+    class T1_RedditObject(@SerializedName("data") val value: RedditCommentDynamicData.T1RedditCommentData = RedditCommentDynamicData.T1RedditCommentData(), @SerializedName("kind") val kind: String = "") : RedditReply()
+    class TM_RedditObject(@SerializedName("data") val value: RedditCommentDynamicData.TMore = RedditCommentDynamicData.TMore(), @SerializedName("kind") val kind: String = "") : RedditReply()
 
     companion object {
-        fun flattenComments(rootNode: RedditReplyListingData): MutableList<Pair<Int, RedditCommentDataType>> {
-            val store: MutableList<Pair<Int, RedditCommentDataType>> = mutableListOf()
+        fun flattenComments(rootNode: RedditReplyListingData): MutableList<Pair<Int, RedditCommentDynamicData>> {
+            val store: MutableList<Pair<Int, RedditCommentDynamicData>> = mutableListOf()
             traverse(rootNode, 0) { level, redditCommentData ->
                 store.add(level to redditCommentData)
             }
             return store
         }
 
-        fun traverse(rootNode: RedditReplyListingData, level: Int, callback: (Int, RedditCommentDataType) -> Unit = { _, _ -> }) {
+        fun traverse(rootNode: RedditReplyListingData, level: Int, callback: (Int, RedditCommentDynamicData) -> Unit = { _, _ -> }) {
             if (rootNode.valueList.isEmpty()) {
                 return
             }
